# docs/openapi/b1.yaml
openapi: 3.1.0
info:
  title: LapanganGo Phase B2 API
  version: 2.0.0
  description: Kontrak kumulatif B1 dan B2 untuk runtime lokal sandbox LapanganGo.
...
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /business/tenants/transfer-primary-owner:
    post:
      tags: [Business]
      operationId: transferPrimaryOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, targetMembershipId, reason]
              properties:
                tenantId: { $ref: "#/components/schemas/PublicId" }
                targetMembershipId: { $ref: "#/components/schemas/PublicId" }
                reason: { type: string, minLength: 5, maxLength: 500 }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "204": { description: Primary Owner transferred transactionally. }
...
  /business/venues:
    get:
      tags: [Business]
      operationId: listBusinessVenues
      parameters:
        - $ref: "#/components/parameters/TenantIdQuery"
      responses:
        "200":
          description: Venues visible in the active workspace.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/VenueSetupSummary" }
    post:
      tags: [Business]
      operationId: createBusinessVenue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, name]
              properties:
                tenantId: { $ref: "#/components/schemas/PublicId" }
                name: { type: string, minLength: 3, maxLength: 80 }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Private venue draft created.
          content:
            application/json:
              schema:
                type: object
                required: [id, slug]
                properties:
                  id: { $ref: "#/components/schemas/PublicId" }
                  slug: { type: string }
  /business/venues/{venueId}:
    get:
      tags: [Business]
      operationId: getBusinessVenue
      parameters:
        - $ref: "#/components/parameters/VenueId"
        - $ref: "#/components/parameters/TenantIdQuery"
      responses:
        "200":
          description: Complete venue setup read model.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VenueSetupDetail" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /business/venues/{venueId}/profile:
    put:
      tags: [Business]
      operationId: updateBusinessVenueProfile
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VenueProfileInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "204": { description: Profile saved. }
  /business/venues/{venueId}/catalog:
    put:
      tags: [Business]
      operationId: updateBusinessVenueCatalog
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, sportIds, facilityIds]
              properties:
                tenantId: { $ref: "#/components/schemas/PublicId" }
                sportIds:
                  type: array
                  minItems: 1
                  items: { $ref: "#/components/schemas/PublicId" }
                facilityIds:
                  type: array
                  items: { $ref: "#/components/schemas/PublicId" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "204": { description: Venue sports and facilities replaced. }
  /business/venues/{venueId}/courts:
    post:
      tags: [Business]
      operationId: createBusinessCourt
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCourtInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Court created with default booking settings.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { $ref: "#/components/schemas/PublicId" }
  /business/venues/{venueId}/courts/{courtId}/availability:
    put:
      tags: [Business]
      operationId: updateCourtAvailability
      parameters:
        - $ref: "#/components/parameters/VenueId"
        - name: courtId
          in: path
          required: true
          schema: { $ref: "#/components/schemas/PublicId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CourtAvailabilityInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "204": { description: Schedule and booking constraints replaced. }
  /business/venues/{venueId}/exceptions:
    post:
      tags: [Business]
      operationId: createScheduleException
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleExceptionInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Exception created.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { $ref: "#/components/schemas/PublicId" }
  /business/venues/{venueId}/payment-settings:
    put:
      tags: [Business]
      operationId: updateVenuePaymentSettings
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VenuePaymentSettingsInput" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "204": { description: Payment settings saved. }
...
  /business/venues/{venueId}/addons:
    post:
      tags: [Business]
      operationId: createVenueAddon
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, name, price]
              properties:
                tenantId: { $ref: "#/components/schemas/PublicId" }
                name: { type: string, minLength: 2, maxLength: 60 }
                price: { type: integer, minimum: 0 }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201": { description: Add-on created. }
  /business/venues/{venueId}/submit:
    post:
      tags: [Business]
      operationId: submitBusinessVenue
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId]
              properties:
                tenantId: { $ref: "#/components/schemas/PublicId" }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "201":
          description: Versioned verification request created.
          content:
            application/json:
              schema:
                type: object
                required: [requestId]
                properties:
                  requestId: { $ref: "#/components/schemas/PublicId" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /admin/verifications/{requestId}/decision:
    post:
      tags: [Admin]
      operationId: decideAdminVerification
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { $ref: "#/components/schemas/PublicId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, reason]
              properties:
                decision:
                  type: string
                  enum: [APPROVED, REJECTED, REVISION_REQUIRED]
                reason: { type: string, minLength: 5, maxLength: 2000 }
      responses:
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "204": { description: Decision recorded. }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  parameters:
...
    TenantIdQuery:
      name: tenantId
      in: query
      required: true
      schema: { $ref: "#/components/schemas/PublicId" }
    VenueId:
      name: venueId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PublicId" }
...
  responses:
...
    Forbidden:
      description: The active identity cannot access this workspace resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Conflict:
      description: State or reservation conflict.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    PayloadTooLarge:
      description: JSON request body exceeds the global 1 MB limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
...
    PublicId:
      type: string
      pattern: "^[A-Za-z0-9_-]{22}$"
      description: Deterministically encrypted reference; internal numeric key is never exposed.
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
        requestId: { type: string }
...
    SetupProgress:
      type: object
      required: [percentage, complete, sections]
      properties:
        percentage: { type: integer, minimum: 0, maximum: 100 }
        complete: { type: boolean }
        sections:
          type: object
          additionalProperties: { type: boolean }
    VenueSetupSummary:
      type: object
      required:
        [id, tenantId, name, slug, addressLine, status, publicationStatus, progress]
      properties:
        id: { $ref: "#/components/schemas/PublicId" }
        tenantId: { $ref: "#/components/schemas/PublicId" }
        name: { type: string }
        slug: { type: string }
        addressLine: { type: string }
        status: { type: string }
        publicationStatus: { type: string }
        progress: { $ref: "#/components/schemas/SetupProgress" }
        revisionReason: { type: [string, "null"] }
    VenueSetupDetail:
      allOf:
        - $ref: "#/components/schemas/VenueSetupSummary"
        - type: object
          required:
            [
              description,
              phoneE164,
              provinceCode,
              cityCode,
              districtCode,
              postalCode,
              timezone,
              indoorOutdoorType,
              houseRules,
              emergencyContact,
              sportIds,
              facilityIds,
              mediaCount,
              courts,
              exceptions,
              blocks,
              priceRules,
              addons,
            ]
          properties:
            description: { type: string }
            phoneE164: { type: string }
            email: { type: [string, "null"] }
            provinceCode: { type: [string, "null"] }
            cityCode: { type: [string, "null"] }
            districtCode: { type: [string, "null"] }
            postalCode: { type: [string, "null"] }
            latitude: { type: [string, "null"] }
            longitude: { type: [string, "null"] }
            timezone: { type: string }
            indoorOutdoorType: { type: string }
            parkingInfo: { type: [string, "null"] }
            houseRules: { type: string }
            emergencyContact: { type: [string, "null"] }
            sportIds:
              type: array
              items: { $ref: "#/components/schemas/PublicId" }
            facilityIds:
              type: array
              items: { $ref: "#/components/schemas/PublicId" }
            mediaCount: { type: integer }
            courts:
              type: array
              items: { $ref: "#/components/schemas/BusinessCourt" }
            exceptions:
              type: array
              items: { type: object, additionalProperties: true }
            blocks:
              type: array
              items: { type: object, additionalProperties: true }
            priceRules:
              type: array
              items: { type: object, additionalProperties: true }
            paymentSettings:
              oneOf:
                - $ref: "#/components/schemas/VenuePaymentSettings"
                - type: "null"
            addons:
              type: array
              items: { type: object, additionalProperties: true }
    BusinessCourt:
      type: object
      required: [id, name, sportId, status, weeklySchedule]
      properties:
        id: { $ref: "#/components/schemas/PublicId" }
        name: { type: string }
        sportId: { $ref: "#/components/schemas/PublicId" }
        surface: { type: [string, "null"] }
        capacity: { type: [integer, "null"] }
        status: { type: string }
        settings: { type: [object, "null"], additionalProperties: true }
        weeklySchedule:
          type: array
          items: { type: object, additionalProperties: true }
...
    VenueProfileInput:
      type: object
      required:
        [
          tenantId,
          name,
          description,
          phoneE164,
          addressLine,
          latitude,
          longitude,
          timezone,
          indoorOutdoorType,
          houseRules,
        ]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        name: { type: string, minLength: 3, maxLength: 80 }
        description: { type: string, minLength: 20, maxLength: 5000 }
        phoneE164: { type: string }
        email: { type: string, format: email }
        addressLine: { type: string, minLength: 10, maxLength: 255 }
        provinceCode: { type: string, maxLength: 10 }
        cityCode: { type: string, maxLength: 10 }
        districtCode: { type: string, maxLength: 10 }
        postalCode: { type: string, pattern: "^\\d{5}$" }
        latitude: { type: string }
        longitude: { type: string }
        timezone: { type: string }
        indoorOutdoorType: { type: string, enum: [INDOOR, OUTDOOR, MIXED] }
        parkingInfo: { type: string, maxLength: 255 }
        houseRules: { type: string, minLength: 10, maxLength: 10000 }
        emergencyContact: { type: string, maxLength: 50 }
    CreateCourtInput:
      type: object
      required: [tenantId, sportId, name]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        sportId: { $ref: "#/components/schemas/PublicId" }
        name: { type: string, minLength: 2, maxLength: 50 }
        surface: { type: string, maxLength: 50 }
        capacity: { type: integer, minimum: 1, maximum: 500 }
    CourtAvailabilityInput:
      type: object
      required:
        [
          tenantId,
          intervalMinutes,
          bufferMinutes,
          minimumDurationMinutes,
          maximumDurationMinutes,
          bookingWindowDays,
          minimumLeadMinutes,
          weeklySchedule,
        ]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        intervalMinutes: { type: integer }
        bufferMinutes: { type: integer }
        minimumDurationMinutes: { type: integer }
        maximumDurationMinutes: { type: integer }
        bookingWindowDays: { type: integer }
        minimumLeadMinutes: { type: integer }
        weeklySchedule:
          type: array
          items:
            type: object
            required: [dayOfWeek, opensAt, closesAt, active]
            properties:
              dayOfWeek: { type: integer, minimum: 0, maximum: 6 }
              opensAt: { type: string }
              closesAt: { type: string }
              active: { type: boolean }
    ScheduleExceptionInput:
      type: object
      required: [tenantId, localDate, kind]
      properties:
        tenantId: { $ref: "#/components/schemas/PublicId" }
        courtId: { $ref: "#/components/schemas/PublicId" }
        localDate: { type: string, format: date }
        kind: { type: string, enum: [OPEN, CLOSED, CUSTOM_HOURS] }
        opensAt: { type: string }
        closesAt: { type: string }
        reason: { type: string, maxLength: 255 }
    VenuePaymentSettings:
      type: object
      required:
        [
          allowFull,
          allowDp,
          dpPercentage,
          allowPayAtVenue,
          reservationAmount,
          manualConfirmationMinutes,
          balanceDeadlineMinutes,
        ]
      properties:
        allowFull: { type: boolean }
        allowDp: { type: boolean }
        dpPercentage: { type: [integer, "null"] }
        allowPayAtVenue: { type: boolean }
        reservationAmount: { type: [integer, "null"] }
        manualConfirmationMinutes: { type: integer }
        balanceDeadlineMinutes: { type: [integer, "null"] }
    VenuePaymentSettingsInput:
      allOf:
        - $ref: "#/components/schemas/VenuePaymentSettings"
        - type: object
          required: [tenantId]
          properties:
            tenantId: { $ref: "#/components/schemas/PublicId" }
...